factor out integration tests
