+++ crates/evm-prover/src/lib.rs
@@ -1,0 +1,8 @@
+pub mod commands;
+pub mod config;
+pub mod grpc;
+pub mod proto;
+pub mod prover;
+pub mod storage;
+#[cfg(test)]
+pub mod tests_storage;
